+++ extensions/1bookmark/CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1bookmark Changelog
+# 1Bookmark Changelog
 
 ## [Space Authentication Policy] - {PR_MERGE_DATE}